backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pulp
google-generativeai
python-multipart
python-dotenv
